feature: tela de gerenciar faltando QR CODE

+++ Server/Controllers/usuario.controller.js
@@ -1,4 +1,5 @@
 import Usuario from "../Models/usuario.model.js";
+import authController from "../Controllers/auth.controller.js";
 
 function findAll(request, response) {
   Usuario
@@ -39,7 +40,7 @@ function create(nome, email, senha, superadmin, response, getToken) {
       nome: nome,
       email: email,
       senha: senha,
-      ehSuperadmin: superadmin
+      ehSuperadmin: false
     })
     .then((result) => {
         
@@ -55,6 +56,23 @@ function create(nome, email, senha, superadmin, response, getToken) {
     });
 }
 
+async function getByToken(request, response){
+    const result = authController.decodeToken(request.params.token);
+    Usuario
+    .findByPk(result.sub)
+    .then(res => {
+      if (res) {
+        response.status(200).send(res);
+      } else {
+        response.status(404).send("Usuário não encontrado");
+      }
+    })
+    .catch(err => {
+      response.status(500).json(err);
+    });
+   
+}
+
 function deleteByPk(request, response) {
   Usuario
     .destroy({ where: { id_usuario: request.params.id } })
@@ -68,6 +86,46 @@ function deleteByPk(request, response) {
     .catch(err => {
       response.status(500).json(err);
     });
+}
+
+async function updateAdmin(request, response) {
+  
+  try {
+      let storedToken = request.params.token;
+      const responsePassword = await authController.testPassword(storedToken, request.body.senha);
+      
+    console.log(responsePassword);
+      if (responsePassword == true){
+          const decodedToken = await authController.decodeToken(storedToken);
+      console.log(decodedToken);
+          const id = decodedToken.sub;
+
+          const hashedPassword = authController.hashSenha(request.body.novaSenha);
+
+          const updatedUsuario = await Usuario.update(
+          {
+            nome: request.body.nome,
+            email: request.body.email,
+            senha: hashedPassword
+          },
+            { where: { id_usuario: id } }
+          );
+
+          if (updatedUsuario[0] > 0) {
+            response.status(200).send({id});
+          } else {
+            response.status(404).send("Usuário não encontrado");
+          }      
+      } 
+      else {
+        console.error("Erro: senha incorreta", response.statusText);
+        response.status(404).send("Senha incorreta");
+      }
+  } catch (err) {
+      console.error("Erro ao alterar dados", err);
+      response.status(500).send("Erro:", err);
+  }
+
 }
 
 function update(request, response) {
@@ -102,4 +160,4 @@ async function validaSuperAdmin(id) {
   }
 }
 
-export default { findAll, findById, create, deleteByPk, update, findByEmail, validaSuperAdmin };
+export default { findAll, findById, create, deleteByPk, update, findByEmail, validaSuperAdmin, updateAdmin, getByToken };
